# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/ingredients/": {
      "get": {
        "tags": [
          "ingredients"
        ],
        "summary": "Get All Ingredients",
        "description": "Return all ingredients.",
        "operationId": "get_all_ingredients_api_ingredients__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngredientRead"
                  },
                  "type": "array",
                  "title": "Response Get All Ingredients Api Ingredients  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ingredients"
        ],
        "summary": "Add Ingredient",
        "description": "Create a new ingredient.",
        "operationId": "add_ingredient_api_ingredients__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredients/possible_tags": {
      "get": {
        "tags": [
          "ingredients"
        ],
        "summary": "Get All Possible Tags",
        "description": "Return all possible ingredient tags ordered by name.",
        "operationId": "get_all_possible_tags_api_ingredients_possible_tags_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PossibleIngredientTag"
                  },
                  "type": "array",
                  "title": "Response Get All Possible Tags Api Ingredients Possible Tags Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ingredients"
        ],
        "summary": "Create Possible Tag",
        "description": "Create a new possible ingredient tag.",
        "operationId": "create_possible_tag_api_ingredients_possible_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PossibleIngredientTagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PossibleIngredientTag"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredients/possible_tags/{tag_id}": {
      "put": {
        "tags": [
          "ingredients"
        ],
        "summary": "Update Possible Tag",
        "description": "Update an existing possible ingredient tag.",
        "operationId": "update_possible_tag_api_ingredients_possible_tags__tag_id__put",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PossibleIngredientTagUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PossibleIngredientTag"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ingredients"
        ],
        "summary": "Delete Possible Tag",
        "description": "Delete a possible ingredient tag if not linked.",
        "operationId": "delete_possible_tag_api_ingredients_possible_tags__tag_id__delete",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Possible Tag Api Ingredients Possible Tags  Tag Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredients/{ingredient_id}": {
      "get": {
        "tags": [
          "ingredients"
        ],
        "summary": "Get Ingredient",
        "description": "Retrieve a single ingredient by ID.",
        "operationId": "get_ingredient_api_ingredients__ingredient_id__get",
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ingredient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ingredients"
        ],
        "summary": "Update Ingredient",
        "description": "Update an existing ingredient.",
        "operationId": "update_ingredient_api_ingredients__ingredient_id__put",
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ingredient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ingredients"
        ],
        "summary": "Delete Ingredient",
        "description": "Delete an ingredient.",
        "operationId": "delete_ingredient_api_ingredients__ingredient_id__delete",
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ingredient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Ingredient Api Ingredients  Ingredient Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngredientCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nutrition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NutritionCreate"
              },
              {
                "type": "null"
              }
            ]
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/IngredientUnitCreate"
            },
            "type": "array",
            "title": "Units"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "IngredientCreate",
        "description": "Schema for creating an ingredient."
      },
      "IngredientRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nutrition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Nutrition"
              },
              {
                "type": "null"
              }
            ]
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/IngredientUnit"
            },
            "type": "array",
            "title": "Units"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/PossibleIngredientTag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "IngredientRead",
        "description": "Schema for reading ingredient data."
      },
      "IngredientUnit": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "ingredient_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredient Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "grams": {
            "type": "number",
            "title": "Grams"
          }
        },
        "type": "object",
        "required": [
          "name",
          "grams"
        ],
        "title": "IngredientUnit",
        "description": "Measurement unit for an ingredient."
      },
      "IngredientUnitCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "grams": {
            "type": "number",
            "title": "Grams"
          }
        },
        "type": "object",
        "required": [
          "name",
          "grams"
        ],
        "title": "IngredientUnitCreate",
        "description": "Schema for creating ingredient unit data."
      },
      "IngredientUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nutrition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NutritionCreate"
              },
              {
                "type": "null"
              }
            ]
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/IngredientUnitCreate"
            },
            "type": "array",
            "title": "Units"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "IngredientUpdate",
        "description": "Schema for updating an ingredient."
      },
...
      "Nutrition": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "ingredient_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredient Id"
          },
          "calories": {
            "type": "number",
            "title": "Calories"
          },
          "fat": {
            "type": "number",
            "title": "Fat"
          },
          "carbohydrates": {
            "type": "number",
            "title": "Carbohydrates"
          },
          "protein": {
            "type": "number",
            "title": "Protein"
          },
          "fiber": {
            "type": "number",
            "title": "Fiber"
          }
        },
        "type": "object",
        "required": [
          "calories",
          "fat",
          "carbohydrates",
          "protein",
          "fiber"
        ],
        "title": "Nutrition",
        "description": "Nutritional information for a single ingredient."
      },
      "NutritionCreate": {
        "properties": {
          "calories": {
            "type": "number",
            "title": "Calories"
          },
          "fat": {
            "type": "number",
            "title": "Fat"
          },
          "carbohydrates": {
            "type": "number",
            "title": "Carbohydrates"
          },
          "protein": {
            "type": "number",
            "title": "Protein"
          },
          "fiber": {
            "type": "number",
            "title": "Fiber"
          }
        },
        "type": "object",
        "required": [
          "calories",
          "fat",
          "carbohydrates",
          "protein",
          "fiber"
        ],
        "title": "NutritionCreate",
        "description": "Schema for creating nutrition data."
      },
      "PossibleIngredientTag": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "group": {
            "type": "string",
            "title": "Group"
          }
        },
        "type": "object",
        "required": [
          "name",
          "group"
        ],
        "title": "PossibleIngredientTag",
        "description": "Tag that can be associated with an ingredient."
      },
      "PossibleIngredientTagCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "group": {
            "type": "string",
            "title": "Group"
          }
        },
        "type": "object",
        "required": [
          "name",
          "group"
        ],
        "title": "PossibleIngredientTagCreate",
        "description": "Schema for creating a possible ingredient tag."
      },
      "PossibleIngredientTagUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          }
        },
        "type": "object",
        "title": "PossibleIngredientTagUpdate",
        "description": "Schema for updating a possible ingredient tag."
      },
...
      "TagRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "TagRef",
        "description": "Reference to an existing tag by ID."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}